--- app/models/connection/adapter/base.rb
@@ -1,7 +1,0 @@
-class Connection::Adapter::Base
-  attr_reader :connection_params
-
-  def initialize(connection_params)
-    @connection_params = connection_params
-  end
-end
