+++ config/application.yml
@@ -9,6 +9,12 @@ default: &default
   DEEP_THOUGHT__AUTH__JWT_SECRET_KEY: <%= SecureRandom.urlsafe_base64(128) %>
   DEEP_THOUGHT__AUTH__EMAIL_DOMAIN_PATTERN: .*
   DEEP_THOUGHT__AUTH__CALLBACK_ENDPOINT: http://deepthought.localhost.com/
+  DEEP_THOUGHT__JOB__RABBIT_URI: amqp://localhost:5672/
+  DEEP_THOUGHT__JOB__CONCURRENCY: 1
+  DEEP_THOUGHT__JOB__QUEUE_NAME: deep_thought_default
+  DEEP_THOUGHT__JOB__MAX_RETRIES: 1
+  DEEP_THOUGHT__JOB__RETRY_TTL: 60000
+  DEEP_THOUGHT__JOB__SLEEP_TIME: 0.2
 test: &test
   <<: *default
   DEEP_THOUGHT__LOG_LEVEL: ERROR
