+++ lib/tasks/consumer.rake
@@ -1,0 +1,8 @@
+require "#{Rails.root}/lib/consumer.rb"
+
+namespace :consumer do
+  desc "RUN Background consumer job"
+  task run: :environment do
+    Consumer.new.execute!
+  end
+end
